Split unit tests for options struct
